Added smooth key press for pong
