+++ project/frontend/public/index.html
@@ -56,8 +56,12 @@
 
 	<!-- //ANNA -->
 	<div id="canvas-container">
-		<canvas id="snake" width="800" height="600" style="border:5px solid #ffffff; margin-top: 20px;"></canvas>
-		<button id="start-button">Start</button>
+		<canvas id="pong" width="800" height="600" style="border:5px solid #000000; margin-top: 20px;"></canvas>
+		<button id="start-button-pong">Start</button>
+	</div>
+	<div id="canvas-container">
+		<canvas id="snake" width="800" height="600" style="border:5px solid #000000; margin-top: 20px;"></canvas>
+		<button id="start-button-snake">Start</button>
 	</div>
 	<!-- //ANNA -->
 	<script src="../dist/main.js" type="module"></script>
